==> codes/1. extractVideoList.ipynb <==
user id로 channel id 추출하기  
api 직접 쏴서 호출 ^0^  



#https://www.googleapis.com/youtube/v3/channels?key={YOUR_API_KEY}&forUsername={USER_NAME}&part=id

기술관련 유튜버 
1. Marques Brownlee: UCBJycsmduvYEL83R_U4JriQ
2. Linus tech tips: UCXuqSBlHAE6Xw-yeJA0Tunw

교육관련 유튜버
1. Vsause: UC6nSFpj9HTCZ5t-N3Rm3-HA

건강관련 유튜버
1. HealthNut Nutrition: UCZB32syI0FFtThd6xkPTRrg
2. POPSUGAR Fitness: UCBINFWq52ShSgUFEoynfSwg
3. Yoga With Adriene: UCFKE7WVJfvaHW5q283SxchA

정치관련 유튜버
1. Ben Shapiro: UCnQC_G5Xsjhp9fEJKuIcrSw
2. Conservative Twins: UCFfD9fWqkT4PNy7v6r8ilmg
3. David Pakman Show: UCvixJtaXuNdMPUGdOPcY8Ag


Extract Video List

import json
from urllib.request import urlopen
import pandas as pd



channel_id = "UCvixJtaXuNdMPUGdOPcY8Ag"
maxrsults = "200"
apikey = "AIzaSyBX9LuZU8g6LtrGaru6D68x6g-k69u3zCo"
#CMgBEAA

url1 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey
url2 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CDIQAA"
url3 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CGQQAA"
url4 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CJYBEAA"
url5 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CMgBEAA"
url6 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CPoBEAA"
url7 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CKwCEAA"
url8 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CN4CEAA"
url9 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CJADEAA"
url10 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CJADEAA"
url11 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CMIDEAA"
url12 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CPQDEAA"
url12 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CPQDEAA"
url12 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CPQDEAA"
url12 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CPQDEAA"
url12 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CPQDEAA"
url12 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CPQDEAA"
url12 = "https://www.googleapis.com/youtube/v3/search?order=date&part=snippet&channelId="+channel_id+"&maxResults="+maxrsults+"&key="+apikey+"&pageToken=CPQDEAA"



ids = []
titles = []
dates = []

def get_response(url):
  response = urlopen(url).read().decode('utf-8')
  responseJson = json.loads(response)
  extract_info(responseJson)

def extract_info(responseJson):
  for k in responseJson['items']:

    try:
      ids.append(k['id']['videoId'])
    except:
      ids.append("err")
      titles.append("err")
      dates.append("err")
      continue

    try:
      titles.append(k['snippet']['title'])
    except:
      titles.append("err")
      dates.append("err")
      continue

    try:
      dates.append(k['snippet']['publishTime'])
    except:
      dates.append("err")
      continue


get_response(url1)
get_response(url2)
get_response(url3)
get_response(url4)
get_response(url5)
get_response(url6)
get_response(url7)
get_response(url8)
get_response(url9)
get_response(url10)
get_response(url11)

df = pd.DataFrame([ids,titles,dates]).T
df.columns = ['ids','titles','Date']
df = df.drop(df[(df['ids'] == 'err') | (df['titles'] == 'err') | (df['Date'] == "err")].index)

df['Date'] = pd.to_datetime(df['Date'], format='%Y-%m-%d %H:%M:%S')

df = df[(df['Date'].dt.year == 2019)|(df['Date'].dt.year == 2020)|(df['Date'].dt.year == 2021)]

df = df.reset_index(drop=True)

for i in range(len(df['ids'])):
  idx = df['ids'][i].find("-")
  if idx==0:
    df.drop([i],axis=0,inplace = True)
df = df.reset_index(drop=True)

df.to_csv('/content/drive/MyDrive/졸프/youtuberdata/DavidPakmanShowVideoLists4.csv',encoding = 'utf-8-sig',index = False)
