==> codes/2. extractCaptionandComment.ipynb <==
!pip install youtube_transcript_api

import pandas as pd
from youtube_transcript_api import YouTubeTranscriptApi
import time
import random

youtuber_info = pd.read_csv("/content/drive/MyDrive/졸프/youtuberdata/DavidPakmanShowVideoLists4.csv",encoding='utf-8-sig',parse_dates=["Date"])
video_list = youtuber_info['ids']
year = youtuber_info['Date'].dt.year
month = youtuber_info['Date'].dt.month
time_code = []

T0: 코로나 이전 2019.01 - 2019.12  
T1: 코로나 초기 2020.01 - 2020.06  
T2: 코로나 중기 2020.07 - 2020.12  
T3: 코로나 후기 2021.01 - Present  


코로나 시기별로 분류하는 코드  


for i in range(len(youtuber_info['Date'])):
  time_code.append("없음")

for i in range(len(youtuber_info['Date'])):
  if year[i] == 2021:
    time_code[i] = "T3"
  elif year[i] == 2020 and month[i]<=12 and month[i]>=7:
    time_code[i] = "T2"
  elif year[i] == 2020 and month[i]<=6 and month[i]>=1:
    time_code[i] = "T1"
  elif year[i] == 2019:
    time_code[i] = "TO"
    

tdf = pd.DataFrame(time_code,columns=['time_code'])
youtuber_info['time_code'] = tdf['time_code']

youtuber_info

만약, 비디오를 랜덤으로 추출해서 분석하고 싶으면 이 코드 사용. 

sample_df = pd.DataFrame(columns = ['ids','titles','Date','time_code']) ## 데이터 초기값
for g in youtuber_info['time_code'].unique():
    temp_df = youtuber_info.query('time_code==@g').sample(n=5) ## 그룹별 데이터 추출 및 2개 비복원 추출
    sample_df = pd.concat([sample_df,temp_df]) ## 데이터 추가
    
sample_df = sample_df.reset_index(drop=True)

#sample_df
sample_df.to_csv("DavidList.csv",encoding='utf-8-sig')

sample_df

자막 추출  

def extract_caption(video_id,time_code):
  srt = YouTubeTranscriptApi.get_transcript(video_id)
  #for i in srt:
    #print(i['text'])
  with open('/content/drive/MyDrive/졸프/youtuberdata/DavidCaptionList/'+time_code+"_"+video_id+'_srt_file'+'.txt','w') as f:
    f.writelines('%s\n' % i['text'] for i in srt)

랜덤 추출 안한 경우

for i in range(len(youtuber_info['ids'])):
  try:
    extract_caption(youtuber_info['ids'][i],youtuber_info['time_code'][i])
    time.sleep(1)
  except:
    print(youtuber_info['ids'][i])

랜덤 추출 한 경우

for i in range(len(sample_df['ids'])):
  try:
    extract_caption(sample_df['ids'][i],sample_df['time_code'][i])
    time.sleep(1)
  except:
    print(sample_df['ids'][i])

댓글 추출  

!pip install google-api-python-client
import pandas
from googleapiclient.discovery import build

 
api_key = 'AIzaSyCCCaJqaOhewWzuOhMjEz8WtyYZlwBKlTs'
 

youtuber_info = pd.read_csv("/content/drive/MyDrive/졸프/youtuberdata/MarquesBrownleeVideoLists4.csv",encoding='utf-8-sig',parse_dates=["Date"])
video_list = youtuber_info['ids']
year = youtuber_info['Date'].dt.year

video_list = youtuber_info['ids']
year = youtuber_info['Date'].dt.year
month = youtuber_info['Date'].dt.month
time_code = []

for i in range(len(youtuber_info['Date'])):
  time_code.append("없음")

for i in range(len(youtuber_info['Date'])):
  if year[i] == 2021:
    time_code[i] = "T3"
  elif year[i] == 2020 and month[i]<=12 and month[i]>=7:
    time_code[i] = "T2"
  elif year[i] == 2020 and month[i]<=6 and month[i]>=1:
    time_code[i] = "T1"
  elif year[i] == 2019:
    time_code[i] = "TO"

tdf = pd.DataFrame(time_code,columns=['time_code'])
youtuber_info['time_code'] = tdf['time_code']

def extract_comment(video_id,time_code):
  comments = list()
  api_obj = build('youtube', 'v3', developerKey=api_key)


  response = api_obj.commentThreads().list(part='snippet,replies', videoId=video_id, maxResults=100).execute()
 
  while response:
      for item in response['items']: 
          comment = item['snippet']['topLevelComment']['snippet']
          comments.append([comment['textDisplay'], comment['authorDisplayName'], comment['publishedAt'], comment['likeCount']])
 
          if item['snippet']['totalReplyCount'] > 0:
              for reply_item in item['replies']['comments']:
                  reply = reply_item['snippet']
                  comments.append([reply['textDisplay'], reply['authorDisplayName'], reply['publishedAt'], reply['likeCount']])
 
      if 'nextPageToken' in response:
          response = api_obj.commentThreads().list(part='snippet,replies', videoId=video_id, pageToken=response['nextPageToken'], maxResults=100).execute()
      else:
          break
 
  file_name = "/content/drive/MyDrive/졸프/youtuberdata/DavidCommentList/"+time_code+"_"+video_id+"_comment.csv"
  df = pandas.DataFrame(comments)
  df.to_csv(file_name, header=['comment', 'id', 'date', 'num_likes'], index=None)

랜덤 추출 안한 경우 

for i in range(len(youtuber_info['ids'])):
  try:
    extract_comment(youtuber_info['ids'][i],youtuber_info['time_code'][i])
  except:
    print(youtuber_info['ids'][i])

랜덤 추출 한 경우

for i in range(len(sample_df['ids'])):
  try:
    extract_comment(sample_df['ids'][i],sample_df['time_code'][i])
  except:
    print(sample_df['ids'][i])
