==> codes/3-3. merge sentiment and emotion analysis result.ipynb <==
3-3. 결과합치기

import pandas as pd
from ast import literal_eval

TravelYoutuber = ['Eva','Kristine','Louis']
Vlogger = ['Zoe','Tess','Nicole']
TechYoutuber = ['Marques','Linus']
HealthYoutuber = ['HealthNut','Popsugar','Yoga']
PoliticsYoutuber = ['Ben','Conservative','David']

youtube_category = 'PoliticsYoutuber'
input_youtubers = PoliticsYoutuber

sentiment 결과 합치기, 유튜버 카테고리별로


capdf = pd.DataFrame(columns = ['youtube_name','time_code','strongPositive','positive','neutral','negative','strongNegative'])
comdf = pd.DataFrame(columns = ['youtube_name','time_code','strongPositive','positive','neutral','negative','strongNegative'])

for input_youtuber in input_youtubers:

  for i in range(4):
    time_code = "T"+str(i)

    input_caption_paths = "/content/drive/MyDrive/졸프/감성분석result/"+input_youtuber+"_Result/T"+str(i)+"_"+input_youtuber+"_Caption_result.csv"
    input_comment_paths = "/content/drive/MyDrive/졸프/감성분석result/"+input_youtuber+"_Result/T"+str(i)+"_"+input_youtuber+"_Comment_result.csv"

    input_caption = pd.read_csv(input_caption_paths,encoding = 'utf-8-sig')
    input_comment = pd.read_csv(input_comment_paths,encoding = 'utf-8-sig')



    cap_sent_val = input_caption['sentiment'].value_counts()
    cap_sent_val_cnt = [cap_sent_val.strongPositive, cap_sent_val.positive, cap_sent_val.neutral, cap_sent_val.negative, cap_sent_val.strongNegative]
    
    com_sent_val = input_comment['sentiment'].value_counts()
    com_sent_val_cnt = [com_sent_val.strongPositive, com_sent_val.positive, com_sent_val.neutral, com_sent_val.negative, com_sent_val.strongNegative]

    capdf=capdf.append({'youtube_name':input_youtuber,'time_code':time_code,'strongPositive':cap_sent_val.strongPositive,
                  'positive':cap_sent_val.positive,'neutral':cap_sent_val.neutral,
                  'negative':cap_sent_val.negative,'strongNegative':cap_sent_val.strongNegative} , ignore_index=True)
    
    comdf=comdf.append({'youtube_name':input_youtuber,'time_code':time_code,'strongPositive':com_sent_val.strongPositive,
                  'positive':com_sent_val.positive,'neutral':com_sent_val.neutral,
                  'negative':com_sent_val.negative,'strongNegative':com_sent_val.strongNegative} , ignore_index=True)
    


for i in range(4):
  sum_cap = capdf[capdf.time_code == 'T'+str(i)].sum()
  sum_com = comdf[comdf.time_code == 'T'+str(i)].sum()

  capdf=capdf.append({'youtube_name':youtube_category,'time_code':'T'+str(i),'strongPositive':sum_cap[2],
                  'positive':sum_cap[3],'neutral':sum_cap[4],
                  'negative':sum_cap[5],'strongNegative':sum_cap[6]} , ignore_index=True)

  comdf=comdf.append({'youtube_name':youtube_category,'time_code':'T'+str(i),'strongPositive':sum_com[2],
                  'positive':sum_com[3],'neutral':sum_com[4],
                  'negative':sum_com[5],'strongNegative':sum_com[6]} , ignore_index=True)



cap_output_path = "/content/drive/MyDrive/졸프/실험결과/sentiment/"+youtube_category+"Result/"+youtube_category+"ResultSentimentCaption.csv"
com_output_path = "/content/drive/MyDrive/졸프/실험결과/sentiment/"+youtube_category+"Result/"+youtube_category+"ResultSentimentComment.csv"

capdf.to_csv(cap_output_path,encoding = 'utf-8-sig',index=False)
comdf.to_csv(com_output_path,encoding = 'utf-8-sig',index=False)

comdf =0
capdf = 0

emotion결과 합치기

def flat_list(array): 
    a=[]
    for i in array:
        if type(i) == type(list()):
            a+=(flat_list(i))       
        else:
            a.append(i)
    return a


capdf = pd.DataFrame(columns = ['youtube_name','time_code','fear','sadness','anger','surprise','anticipation','joy','trust'])
comdf = pd.DataFrame(columns = ['youtube_name','time_code','fear','sadness','anger','surprise','anticipation','joy','trust'])

for input_youtuber in input_youtubers:

  for i in range(4):
    time_code = "T"+str(i)

    input_caption_paths = "/content/drive/MyDrive/졸프/감성분석result/"+input_youtuber+"_Result/T"+str(i)+"_"+input_youtuber+"_Caption_result.csv"
    input_comment_paths = "/content/drive/MyDrive/졸프/감성분석result/"+input_youtuber+"_Result/T"+str(i)+"_"+input_youtuber+"_Comment_result.csv"

    input_caption = pd.read_csv(input_caption_paths, encoding = 'utf-8-sig',converters={'emotion_list':literal_eval})
    input_comment = pd.read_csv(input_comment_paths, encoding = 'utf-8-sig',converters={'emotion_list':literal_eval})

    caption_emotion_list = input_caption['emotion_list'].values.tolist()
    comment_emotion_list = input_comment['emotion_list'].values.tolist()

    flat_caption_emotion_list = flat_list(caption_emotion_list)  
    flat_comment_emotion_list = flat_list(comment_emotion_list)

    fcapdf = pd.DataFrame(flat_caption_emotion_list,columns = ['emotion'])
    fcomdf = pd.DataFrame(flat_comment_emotion_list,columns = ['emotion'])

    cap_emo_val = fcapdf['emotion'].value_counts()
    com_emo_val = fcomdf['emotion'].value_counts()
  
    capdf=capdf.append({'youtube_name':input_youtuber,'time_code':time_code,'fear':cap_emo_val.fear,
                  'sadness':cap_emo_val.sadness,'anger':cap_emo_val.anger,
                  'surprise':cap_emo_val.surprise,
                  'anticipation':cap_emo_val.anticipation,'joy':cap_emo_val.joy,'trust':cap_emo_val.trust,
                  } , ignore_index=True)
    
    comdf=comdf.append({'youtube_name':input_youtuber,'time_code':time_code,'fear':com_emo_val.fear,
                  'sadness':com_emo_val.sadness,'anger':com_emo_val.anger,
                  'surprise':com_emo_val.surprise,
                  'anticipation':com_emo_val.anticipation,'joy':com_emo_val.joy,'trust':com_emo_val.trust,
                  } , ignore_index=True)
    


for i in range(4):
  sum_cap = capdf[capdf.time_code == 'T'+str(i)].sum()
  sum_com = comdf[comdf.time_code == 'T'+str(i)].sum()

  capdf=capdf.append({'youtube_name':youtube_category,'time_code':'T'+str(i),'fear':sum_cap[2],
                  'sadness':sum_cap[3],'anger':sum_cap[4],
                  'surprise':sum_cap[5],'anticipation':sum_cap[6],
                  'joy':sum_cap[7],'trust':sum_cap[8],
                  } , ignore_index=True)

  comdf=comdf.append({'youtube_name':youtube_category,'time_code':'T'+str(i),'fear':sum_com[2],
                  'sadness':sum_com[3],'anger':sum_com[4],
                  'surprise':sum_com[5], 'anticipation':sum_com[6],
                  'joy':sum_com[7],'trust':sum_com[8],
                  } , ignore_index=True)

cap_output_path = "/content/drive/MyDrive/졸프/실험결과/emotion/"+youtube_category+"Result/"+youtube_category+"ResultEmotionCaption.csv"
com_output_path = "/content/drive/MyDrive/졸프/실험결과/emotion/"+youtube_category+"Result/"+youtube_category+"ResultEmotionComment.csv"

capdf.to_csv(cap_output_path,encoding = 'utf-8-sig',index=False)
comdf.to_csv(com_output_path,encoding = 'utf-8-sig',index=False)
