==> codes/4. visualize the results.ipynb <==
라이브러리 임포트

import pandas as pd
from wordcloud import WordCloud, STOPWORDS
import matplotlib.pyplot as plt
import numpy as np
from PIL import Image
%matplotlib inline
from ast import literal_eval
import matplotlib.patches as mpatches

TravelYoutuber = ['Eva','Kristine','Louis']
Vlogger = ['Zoe','Tess','Nicole']
TechYoutuber = ['Marques','Linus']
HealthYoutuber = ['HealthNut','Popsugar','Yoga']
PoliticsYoutuber = ['Ben','Conservative','David']

youtube_category = 'PoliticsYoutuber'

input_youtubers = PoliticsYoutuber

1.sentiment - caption

cap_input_path = "/content/drive/MyDrive/졸프/실험결과/sentiment/"+youtube_category+"Result/"+youtube_category+"ResultSentimentCaption.csv"
capdf = pd.read_csv(cap_input_path, encoding = 'utf-8-sig')

output_path = "/content/drive/MyDrive/졸프/실험결과/sentiment/"+youtube_category+"Result/"

capdf

df = capdf.drop(columns=['youtube_name'], axis=1)
youtube_name_list = capdf['youtube_name']
df = df.set_index('time_code','youtube_name')
res = df.div(df.sum(axis=1), axis=0)

perdf = res.reset_index()

perdf

#그.... 숫자들 안겹치게 보이기
colors = ['#A7CF04','#DCF967','#FBFFC2','#FBB0B0','#F47070']
names =input_youtubers+[youtube_category]

for i in range(len(names)):
  ax = perdf.iloc[i*4:i*4+4].plot(kind='barh', stacked=True, title= names[i]+ " statistics", rot=0,color=colors)
  for p in ax.patches:
      left, bottom, width, height = p.get_bbox().bounds
      ax.annotate("%.1f"%(width*100), xy=(left+width/2, bottom+height/2), ha='center', va='center')

  labels = ['strongPositive', 'positive', 'neutral','negative','strongNegative']
  plt.yticks([0,1,2,3],['T0','T1','T2','T3'])
  plt.box(False)
  lgd = plt.legend(bbox_to_anchor=(1.05, 1),loc='upper left')
  plt.tight_layout()
  plt.savefig(output_path+names[i]+"_BarChartResultSentimentCaption.png",dpi=200,bbox_extra_artist=[lgd])
  plt.show()
  
  print()


1.sentiment-comment

com_input_path = "/content/drive/MyDrive/졸프/실험결과/sentiment/"+youtube_category+"Result/"+youtube_category+"ResultSentimentComment.csv"
comdf = pd.read_csv(com_input_path, encoding = 'utf-8-sig')

output_path = "/content/drive/MyDrive/졸프/실험결과/sentiment/"+youtube_category+"Result/"

comdf

df = comdf.drop(columns=['youtube_name'], axis=1)
youtube_name_list = comdf['youtube_name']
df = df.set_index('time_code','youtube_name')
res = df.div(df.sum(axis=1), axis=0)

perdf = res.reset_index()

perdf

colors = ['#A7CF04','#DCF967','#FBFFC2','#FBB0B0','#F47070']
names =input_youtubers+[youtube_category]

for i in range(len(names)):
  ax = perdf.iloc[i*4:i*4+4].plot(kind='barh', stacked=True, title= names[i]+ " statistics", rot=0,color=colors)
  for p in ax.patches:
      left, bottom, width, height = p.get_bbox().bounds
      ax.annotate("%.1f"%(width*100), xy=(left+width/2, bottom+height/2), ha='center', va='center')

  labels = ['strongPositive', 'positive', 'neutral','negative','strongNegative']
  plt.yticks([0,1,2,3],['T0','T1','T2','T3'])
  plt.box(False)
  lgd = plt.legend(bbox_to_anchor=(1.05, 1),loc='upper left')
  plt.tight_layout()
  plt.savefig(output_path+names[i]+"_BarChartResultSentimentComment.png",dpi=200,bbox_extra_artist=[lgd])
  #plt.legend(pie[0],labels, bbox_to_anchor=(1,0.5), loc="center right", fontsize=10, bbox_transform=plt.gcf().transFigure)
  plt.show()
  
  print()


2. emotion - caption

cap_input_path = "/content/drive/MyDrive/졸프/실험결과/emotion/"+youtube_category+"Result/"+youtube_category+"ResultEmotionCaption.csv"
capdf = pd.read_csv(cap_input_path, encoding = 'utf-8-sig')

output_path = "/content/drive/MyDrive/졸프/실험결과/emotion/"+youtube_category+"Result/"

capdf

df = capdf.drop(columns=['youtube_name'], axis=1)
youtube_name_list = capdf['youtube_name']
df = df.set_index('time_code','youtube_name')
res = df.div(df.sum(axis=1), axis=0)

perdf = res.reset_index()

perdf

colors = ['#656364','#467BBF','#EB503A','#94C4B9','#4ABAAA','#EED83F','#D8D8D8']
names =input_youtubers+[youtube_category]

for i in range(len(names)):
  ax = perdf.iloc[i*4:i*4+4].plot(kind='barh', stacked=True, title= names[i]+ " statistics", rot=0,color=colors)
  for p in ax.patches:
      left, bottom, width, height = p.get_bbox().bounds
      ax.annotate("%.1f"%(width*100), xy=(left+width/2, bottom+height/2), ha='center', va='center')

  labels = ['fear', 'sadness','anger','anticipation','joy','trust','surprise']
  plt.yticks([0,1,2,3],['T0','T1','T2','T3'])
  plt.box(False)
  lgd = plt.legend(bbox_to_anchor=(1.05, 1),loc='upper left')
  plt.tight_layout()
  plt.savefig(output_path+names[i]+"_BarChartResultEmotionCaption.png",dpi=200,bbox_extra_artist=[lgd])
  plt.show()
  
  print()


2. emotion - comment  

com_input_path = "/content/drive/MyDrive/졸프/실험결과/emotion/"+youtube_category+"Result/"+youtube_category+"ResultEmotionComment.csv"
comdf = pd.read_csv(com_input_path, encoding = 'utf-8-sig')

output_path = "/content/drive/MyDrive/졸프/실험결과/emotion/"+youtube_category+"Result/"

comdf

df = comdf.drop(columns=['youtube_name'], axis=1)
youtube_name_list = comdf['youtube_name']
df = df.set_index('time_code','youtube_name')
res = df.div(df.sum(axis=1), axis=0)

perdf = res.reset_index()


perdf

colors = ['#656364','#467BBF','#EB503A','#94C4B9','#4ABAAA','#EED83F','#D8D8D8']
names =input_youtubers+[youtube_category]

for i in range(len(names)):
  ax = perdf.iloc[i*4:i*4+4].plot(kind='barh', stacked=True, title= names[i]+ " statistics", rot=0,color=colors)
  for p in ax.patches:
      left, bottom, width, height = p.get_bbox().bounds
      ax.annotate("%.1f"%(width*100), xy=(left+width/2, bottom+height/2), ha='center', va='center')

  labels = ['fear', 'sadness','anger','anticipation','joy','trust','surprise']
  plt.yticks([0,1,2,3],['T0','T1','T2','T3'])
  plt.box(False)
  plt.legend(bbox_to_anchor=(1.05, 1),loc='upper left')
  plt.tight_layout()
  plt.savefig(output_path+names[i]+"_BarChartResultEmotionComment.png",dpi=200,bbox_extra_artist=[lgd])
  plt.show()
  print()
