==> codes/5. statistics.ipynb <==
import pandas as pd
import os

TravelYoutuber = ['Eva','Kristine','Louis']
Vlogger = ['Zoe','Tess','Nicole']
TechYoutuber = ['Marques','Linus']
HealthYoutuber = ['HealthNut','Popsugar','Yoga']
PoliticsYoutuber = ['Ben','Conservative','David']

def load_files(name,time_code):

  path = '/content/drive/MyDrive/졸프/youtuberdata/'+name+'CommentList/'
  

  files = []

  for r, d, f in os.walk(path):
      for file in f:
          if '.csv' and time_code in str(file):
            
            files.append(os.path.join(r, file))
  
  return files
  

def load_caption_files(name, time_code):

  path = '/content/drive/MyDrive/졸프/감성분석result/'+name+'_Result/'

  files = []

  for r, d, f in os.walk(path):
      for file in f:
          if '.csv' and time_code in str(file):           
            files.append(os.path.join(r, file))
  
  return files

youtuber_name_list = []
time_code_list = []
sum_videoid_list = []
sum_comment_list = []
sum_caption_list = []

youtubers = PoliticsYoutuber
print(youtubers)
for youtuber in youtubers:
  for i in range(4):
    if i == 0:
      time_code1 = 'TO'
    else:
      time_code1 = 'T'+str(i)
    paths = load_files(youtuber, time_code1)
    time_code2 = 'T'+str(i)
    caption_paths = load_caption_files(youtuber,time_code2)

    print(youtuber)
    print(paths)
    sum_videoid = 0
    sum_comment = 0
    sum_caption = 0
    for path in paths:
      sum_videoid += 1
      youtube_data = pd.read_csv(path,encoding = 'utf-8-sig', engine='python')
      sum_comment += len(youtube_data['comment'])
    for caption_path in caption_paths:
      caption_data = pd.read_csv(caption_path,encoding = 'utf-8-sig', engine='python')
      sum_caption += len(caption_data['raw_comment'])

    youtuber_name_list.append(youtuber)
    time_code_list.append('T'+str(i))
    sum_videoid_list.append(sum_videoid)
    sum_comment_list.append(sum_comment)
    sum_caption_list.append(sum_caption)
        
  

df = pd.DataFrame({'youtuber_name':youtuber_name_list, 'time_code':time_code_list,
                   'num_video':sum_videoid_list,'sum_caption':sum_caption_list,'sum_comment':sum_comment_list})

df

for i in range(4):
  sum_com = df[df.time_code == 'T'+str(i)].sum()

  df=df.append({'youtuber_name':'PoliticsYoutuber','time_code':'T'+str(i),'num_video':sum_com[2],
                  'sum_caption':sum_com[3],'sum_comment':sum_com[4]},ignore_index=True)

df

df.to_csv("/content/drive/MyDrive/졸프/실험결과/whole_statistics/PoliticsYoutuber_statistics.csv",encoding = 'utf-8-sig',index = False)
